Fix URL formatting and add URLDATE
Ensured < and > prefix/suffix are formatted correctly in URLs.

Also added urldate tag for last accessed "viewed " element as explained in http://www.lib.uts.edu.au/help/referencing/harvard-uts-referencing-guide/website-and-social-media

The chosen tag (urldate) and it's format (yyyy-mm-dd) was chosen for compatibility with biblatex.

Unfortunately this has necessitated the use of a macro from the datetime2 package in the .bst file. Another alternative is removing this and requiring users to preformat the date or use the macro in their .bib file.